pyproject.toml
[project]
name = "car-appointment-assistant"
version = "0.1.0"
description = "Voice-based car dealership appointment assistant MVP"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-groq>=0.2.0",
    "strawberry-graphql[fastapi]>=0.252.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
start = "main:run"
